Automatic radio parsing!
The radio module now automatically grabs lists of files from /radio folders which contain .wav, .ogg, or .wav files. It also gets the station name from station.py (or the folder name). Station order on the Pip-Boy can be controlled via a number in the front of the folder name.